requirements.txt
customtkinter
pyrebase4
firebase-admin
Pillow
python-dotenv
playsound
plyer
pytest
pre-commit
